Add support for detecting whether the server requires client authetication [#564]
This change adds a new 'client_cert_required' value to the response for the ctls, ztls, and openssl backends that returns the following values:
- true: if the target requires client authentication
- false: if the target does not require client authentication
- null: if we cannot determine whether or not the remote target requires client authentication

This is implemented by checking for TLS alert message 42 (`bad_certificate`) or alert message 116 (`certificate_required`).

In `openssl`, this is done by checking the stderr returned from execOpenSSL() to determine if error messages related to those specific TLS alerts are returned.

In `crypto/tls` and `ztls`, this is accomplished by parsing the underlying error returned in the `net.OpError` returned by the client when a TLS connection fails and searching for the corresponding `bad certificate` and `certificate required` messages. It should be noted that we have explicitly excluded TLS 1.3 here because `crypto/tls` consider the handshake complete and returns from `conn.HandshakeContext()` before it has dealt with the `CertificateRequest` message sent by the server. As a result, the errors are not surfaced clientside until the underlying connection is actually used. In these cases (TLS 1.3), we simply return a `nil` to indicate that we're unable to determine whether or not client authentication is required.